pyproject.toml
[tool.poetry]
name = "llm_atc"
version = "0.1.0"
description = ""
authors = ["Andrew Aikawa <asai@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
omegaconf = "^2.3.0"
posthog = "^3.0.1"

[tool.poetry.scripts]
llm-atc = "llm_atc.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
